Add an index on document_type for content_items
This is used in a query to count the number of content items and
the current 3-way index can't be used as the other fields aren't
present in the query. It takes the running time down from 15
seconds to under 500 ms. It also stops the application from
timing out the first time you go to it.